Renamespaces Acquia\Platform\Cloud\Api\SDK to Acquia\Platform\Cloud\Hosting
- Renamespaces Acquia\Platform\Cloud\Api\SDK\Domain\Domain to Acquia\Platform\Hosting\Domain\Domain
- Renamespaces Acquia\Platform\Cloud\Api\SDK\Environment\Environment to Acquia\Platform\Hosting\Environment\Environment
- Renamespaces Acquia\Platform\Cloud\Api\SDK\Server\Server to Acquia\Platform\Hosting\Server\Server
- Renamespaces Acquia\Platform\Cloud\Api\SDK\Site to Acquia\Platform\Cloud\Hosting\Site\Site
- Renamespaces Acquia\Platform\Cloud\Api\SDK\Sites to Acquia\Platform\Cloud\Hosting\Site\Sites
- Renamespaces Acquia\Platform\Cloud\Api\SDK\Task\Task to Acquia\Platform\Hosting\Task\Task
- Note: Site was placed in a sub-namespace for consistency with the rest.
