+++ src/Hosting/Domain/Domain.php
@@ -4,7 +4,7 @@
  * Contains Domain.php.
  */
 
-namespace Acquia\Platform\Cloud\Api\SDK\Domain;
+namespace Acquia\Platform\Cloud\Hosting\Domain;
 
 
 use Acquia\Platform\Cloud\Api\ClientInterface;
