+++ src/Hosting/Environment/EnvironmentInterface.php
@@ -4,7 +4,7 @@
  * Contains EnvironmentInterface.php.
  */
 
-namespace Acquia\Platform\Cloud\Api\SDK\Environment;
+namespace Acquia\Platform\Cloud\Hosting\Environment;
 
 use Acquia\Platform\Cloud\Api\DataInterface;
 
@@ -51,17 +51,17 @@ interface EnvironmentInterface extends DataInterface {
   public function getLogStream();
 
   /**
-   * @return \Acquia\Platform\Cloud\Api\SDK\Task\TaskInterface
+   * @return \Acquia\Platform\Cloud\Hosting\Task\TaskInterface
    */
   public function enableLiveDev();
 
   /**
-   * @return \Acquia\Platform\Cloud\Api\SDK\Task\TaskInterface
+   * @return \Acquia\Platform\Cloud\Hosting\Task\TaskInterface
    */
   public function disableLiveDev();
 
   /**
-   * @return \Acquia\Platform\Cloud\Api\SDK\Server\ServerInterface[]
+   * @return \Acquia\Platform\Cloud\Hosting\Server\ServerInterface[]
    */
   public function getServers();
 
@@ -69,7 +69,7 @@ interface EnvironmentInterface extends DataInterface {
    * @param $name
    *   The name of the server you wish to retrieve.
    *
-   * @return \Acquia\Platform\Cloud\Api\SDK\Server\ServerInterface
+   * @return \Acquia\Platform\Cloud\Hosting\Server\ServerInterface
    */
   public function getServer($name);
 
@@ -80,25 +80,25 @@ interface EnvironmentInterface extends DataInterface {
 
   /**
    * @param $domain
-   * @return \Acquia\Platform\Cloud\Api\SDK\Domain\DomainInterface
+   * @return \Acquia\Platform\Cloud\Hosting\Domain\DomainInterface
    */
   public function getDomain($domain);
 
   /**
    * @param $domain
-   * @return \Acquia\Platform\Cloud\Api\SDK\Task\TaskInterface
+   * @return \Acquia\Platform\Cloud\Hosting\Task\TaskInterface
    */
   public function createDomain($domain);
 
   /**
    * @param $source
-   * @return \Acquia\Platform\Cloud\Api\SDK\Task\TaskInterface
+   * @return \Acquia\Platform\Cloud\Hosting\Task\TaskInterface
    */
   public function installFromSource($source);
 
   /**
    * @param $make_file
-   * @return \Acquia\Platform\Cloud\Api\SDK\Task\TaskInterface
+   * @return \Acquia\Platform\Cloud\Hosting\Task\TaskInterface
    */
   public function installFromManifest($make_file);
