+++ src/Hosting/Server/Server.php
@@ -4,7 +4,7 @@
  * Contains Server.php.
  */
 
-namespace Acquia\Platform\Cloud\Api\SDK\Server;
+namespace Acquia\Platform\Cloud\Hosting\Server;
 
 
 use Acquia\Platform\Cloud\Api\ClientInterface;
