+++ src/Hosting/Task/Task.php
@@ -1,10 +1,10 @@
 <?php
 /**
  * @file
- * Contains \Acquia\Platform\Cloud\Api\SDK\Task\Task.
+ * Contains \Acquia\Platform\Cloud\Hosting\Task\Task.
  */
 
-namespace Acquia\Platform\Cloud\Api\SDK\Task;
+namespace Acquia\Platform\Cloud\Hosting\Task;
 
 class Task implements TaskInterface {
